Handle consul scrape errors gracefully
A consul scrape error is bad and very likely means consul is not running
under the provided path. While this is certainly something which needs
operator intervention, it's possible to detect that without
crashlooping via the absence of exported metrics or even through an
increase of error responses.

In this setup of running both a consul agent as well as a glimpse-agent
in parallel on a node, it's not necessary to bring down the
glimpse-agent just because consul is not available. This ensures the
availability of service discovery in case consul is still reachable via
the HTTP interface, or at least continues to track statistics about
incoming requests. I believe the worst thing we can do is flying blind
during an outage, therefore prevent a termination if possible in
exchange for the possibility of losing some metrics.