Add support for DNS NS queries
In order to support recursive DNS resolution, this change adds support
for NS queries. It's expected that all servers are registered under one
common domain. For example under srv.example.com. Now it'll be possible
to query any of the servers for the responsible name server for
<zone>.srv.example.com.

Please note that doing such a NS query against a normal agent which is
not a server won't return any results.