Pass full consul info command
In order to allow setups with custom RPC ports, the -rpc flag needs to
be set on the consul info command invocation. Therefore, just expect the
full command to be executed.

Ideally the whole forking will be removed in favor of fetching stats
over the HTTP API (not implemented yet) or at least via RPC call.