Improve reliability of setup in acceptance test
So far we relied on log output of the binary we want to setup. As log
output can arrive "out-of-order" it will be observed as breaking tests.
To make our acceptance test suite deterministic we pass a function with
more involved logic per binary to make the decision if it's ready.
Currently that means we make successfull DNS query to the glimpse-agent
and consul agent is responding with the correct set of instances for a
service created from our test data.