Add store logging
As a separation of concerns logging is realised via an implementation of
the store interface and can be used to wrap logging on top of other
store implementations. It works in a similiar fashion to the
metricStore.
In its current implementation we only log errors as a coarser request
logging on the edge (DNS, HTTP) should be sufficient for any other
purposes. We also introduce a log.Logger which is passed.