Handle requests with unknown zone correctly
So far, when glimpse-agent processes a request for an unknown zone, it
would just log the Consul API error (status code 500) and respond with a
DNS SERVFAIL response.

A SERVFAIL response will cause recursors like unbound to retry the
question a couple of times unnecessarily. The desired behavior is that
clients receive a NXDOMAIN response if the zone is unknown.

With this change, there should be no more errConsulAPI happen during
normal operation, allowing to set up alerts if the error rate goes above
zero.